+++ test/run_tests.sh
@@ -236,6 +236,7 @@ function run_xla_op_tests3 {
   run_device_detection_test "$CDIR/test_gpu_device_detection.py"
   # NOTE: this line below is testing export and don't care about GPU
   PJRT_DEVICE=CPU CPU_NUM_DEVICES=1 run_coverage "$CDIR/test_core_aten_ops.py"
+  run_test "$CDIR/test_pallas.py "
 }
 
 #######################################################################################
